requirements.txt
gitdb==4.0.11
GitPython==3.1.43
smmap==5.0.1
tabulate==0.9.0
tqdm==4.66.5
